Ricardo6

+++ MovEasy/Model/UsuarioModel.cs
@@ -1,6 +1,6 @@
-﻿using ConsoleApp1.Entity;
-using ConsoleApp1.Helpers;
-using ConsoleApp1.Interface;
+﻿using MovEasy.Entity;
+using MovEasy.Helpers;
+using MovEasy.Interface;
 using Dapper;
 using MySql.Data.MySqlClient;
 using System;
@@ -9,21 +9,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ConsoleApp1.Model
+namespace MovEasy.Model
 {
     internal class UsuarioModel : Database, ICrud
     {
         public static void Setup()
         {
-            //using (MySqlConnection con = new MySqlConnection(GetConnectionString()))
-            //{
-            //    string sql = $"DROP TABLE IF EXISTS {UsuarioEntity.DatabaseName}";
-            //    con.Execute(sql);
-            //}
-
-            using (MySqlConnection con = new MySqlConnection(GetConnectionString()))
-            {
-                string sql = $@"CREATE TABLE IF NOT EXISTS {UsuarioEntity.DatabaseName} ( 
+            SQLExecute(
+                $@"CREATE TABLE IF NOT EXISTS {UsuarioEntity.DatabaseName} ( 
                         ID INT NOT NULL AUTO_INCREMENT,
                         DOCUMENTO VARCHAR(45) NOT NULL,
                         TELEFONE1 VARCHAR(11) NOT NULL,
@@ -36,9 +29,8 @@ namespace ConsoleApp1.Model
                         CNH VARCHAR(45),
                         FOTO VARCHAR(255),
                         PRIMARY KEY (ID)
-                )";
-                con.Execute(sql);
-            }
+                )"
+            );
         }
 
         public static void Create()
@@ -66,20 +58,22 @@ namespace ConsoleApp1.Model
                 MostrarUsuarios();
 
                 int ID = GetID("Digite o ID para atualizar: ");
-
-                UsuarioEntity user;
+                bool isLoginValid = VerifyUser(ID);
+                if (!isLoginValid) return;
 
                 string sql = $"SELECT * FROM {UsuarioEntity.DatabaseName} WHERE ID = @ID";
                 var parameters = new { ID };
 
-                user = SQLConnection().QueryFirst<UsuarioEntity>(sql, parameters);
-
+                UsuarioEntity user = SQLConnection().QueryFirst<UsuarioEntity>(sql, parameters);
+                    
                 Console.Clear();
-                Console.WriteLine($"Digite o novo email <{user.EMAIL}>");
-                user.EMAIL = Console.ReadLine();
-
-                Console.WriteLine($"Digite uma nova senha.");
-                user.PASSWORDHASH = PasswordHasher.HashPassword(Console.ReadLine());
+                user.EMAIL = Menu.GetInput($"Digite o novo email <{user.EMAIL}>");
+                user.PASSWORDHASH = PasswordHasher.HashPassword(Menu.GetInput($"Digite uma nova senha."));
+                user.NOME = Menu.GetInput("Digite seu novo Nome");
+                user.SOBRENOME = Menu.GetInput("Digite seu novo Sobrenome");
+                user.DOCUMENTO = Menu.GetInput("Digite seu novo Documento");
+                user.TELEFONE1 = Menu.GetInput("Digite seu novo Telefone");
+                user.TIPO = Convert.ToInt32(Menu.GetInput("Digite seu perfil\n1-Contratante\n2-Motorista"));
 
                 sql = $"UPDATE {UsuarioEntity.DatabaseName} SET {UsuarioEntity.DatabaseValues} WHERE ID = @ID";
 
@@ -103,6 +97,8 @@ namespace ConsoleApp1.Model
                 MostrarUsuarios();
 
                 int ID = GetID("Digite o ID para exclusão: ");
+                bool isLoginValid = VerifyUser(ID);
+                if (!isLoginValid) return;
 
                 SQLExecute(
                     $"DELETE FROM {UsuarioEntity.DatabaseName} WHERE ID = @ID",
@@ -117,8 +113,27 @@ namespace ConsoleApp1.Model
             }
         }
 
+        private static bool VerifyUser(int ID)
+        {
+            string sql = $"SELECT * FROM {UsuarioEntity.DatabaseName} WHERE ID = @ID";
+            var parameters = new { ID };
+
+            UsuarioEntity user = SQLConnection().QueryFirst<UsuarioEntity>(sql, parameters);
+
+            bool login = PasswordHasher.VerifyPassword(user.EMAIL, Menu.GetInput("Digite sua senha."));
+
+            if (login) return true;
+
+            Console.Clear();
+            Console.WriteLine("Senha Inválida! Pressione uma tecla para continuar.");
+            Console.ReadLine();
+            return false;
+            
+        }
+
         public static UsuarioEntity EditUser(UsuarioEntity user)
         {
+            Console.Clear();
             user.EMAIL = Menu.GetInput("Digite seu email");
             user.SENHA = Menu.GetInput("Digite sua Senha");
             user.NOME = Menu.GetInput("Digite seu Nome");
@@ -147,15 +162,16 @@ namespace ConsoleApp1.Model
 
         public static int MostrarUsuarios()
         {
-            string sql = "SELECT * FROM " + UsuarioEntity.DatabaseName;
+            string sql = $"SELECT * FROM {UsuarioEntity.DatabaseName}";
 
             IEnumerable<UsuarioEntity> users = SQLConnection().Query<UsuarioEntity>(sql);
 
+            Console.Clear();
             foreach (UsuarioEntity user in users)
             {
                 Console.WriteLine($"[{user.ID}] {user.NOME} - {user.EMAIL} - {user.TELEFONE1} - {user.DOCUMENTO}");
             }
-
+            Menu.DivisoriaMenu();
             return users.Count();
         }
